% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spSeudoMap_R.R
\name{pred_cellf_spSeudoMap}
\alias{pred_cellf_spSeudoMap}
\title{Function to implement spSeudoMap in python
cell type mapping of spatial transcriptomics using unmatched single-cell RNA-seq data}
\usage{
pred_cellf_spSeudoMap(
  sp_data,
  sc_data,
  outdir = ".",
  sp_subset = FALSE,
  spot.cluster.name = "seurat_clusters",
  spot.cluster.of.interest = NULL,
  env.select = "virtual",
  python.install = F,
  python_path = NULL,
  env.name = "spSeudoMap",
  gpu = TRUE,
  metadata_celltype = "celltype",
  num_markers = 10,
  mixture_mode = "pseudotype",
  seed_num = 0,
  mk_ratio_fix = T,
  mk_ratio = 2,
  pseudo_num_genes = 40,
  pseudo_frac_m = 0.5,
  pseudo_frac_std = 0.05,
  num_top_genes = 20,
  nmix = 10,
  npseudo = 60000,
  alpha = 1,
  alpha_lr = 5,
  emb_dim = 64,
  batch_size = 512,
  n_iterations = 3000,
  init_train_epoch = 10
)
}
\arguments{
\item{outdir}{the directory to save output files (models and results) (default = '.')}

\item{sp_subset}{whether to subset spatial data and calculate for specific spot cluster (default = FALSE)}

\item{spot.cluster.name}{group name of the cluster used for subsetting spatial data (default: 'seurat_clusters')}

\item{spot.cluster.of.interest}{name of each spot clusters to be used (default: NULL)}

\item{env.select}{select between using reticulate virtual environment or conda environment ("virtual" or "conda")}

\item{python.install}{whether to automatically install python version 3.7.12}

\item{python_path}{path for the python 3.7. (default: NULL)
\itemize{
\item If NULL, python version 3.7.12 will be installed (valid for Linux)
\item If "current", python interpreter associated with current virtual env (ex: r-reticulate) will be used. (version should be 3.7)
}}

\item{env.name}{name of the virtual or conda environment to use for CellDART analysis (default: 'spSeudoMap')}

\item{gpu}{check whether to use gpu (True) or not (False) (default = True)}

\item{metadata_celltype}{column name for single-cell annotation data in metadata (default: 'celltype')}

\item{num_markers}{number of selected marker genes in each cell-type (default = 20)}

\item{mixture_mode}{mode of the pseudospot generation: 'default'-when cell types are similar between single-cell/spatial data (Identical to CellDART)
\itemize{
\item 'pseudotype': when there is a mismatch of cell types between single-cell/spatial data and cell types exclusively present in spatial data is considered when generating pseudospots
}}

\item{seed_num}{seed to be used in random sampling (default = 0)}

\item{pseudo_frac_m}{average of presumed fraction of the pseudotype (cell types exclusively present in spatial data) across all spatial spots (default = 0.5)
\itemize{
\item determined by cell sorting study or by literature based evidence that enables speculation of average fraction in the tissue
}}

\item{pseudo_frac_std}{standard deviation of the distribution of presumed pseudotype fraction across all spatial spots (default = 0.1)}

\item{num_top_genes}{number of top genes having highest log fold change between spatial and single-cell normalized pseudobulk counts (spatial/single-cell)
\itemize{
\item the top genes are used for generating module score to predict pseudotype fraction in spatial spots
}}

\item{nmix}{the number of cells sampled from single-cell data when making a pseudospot (default = 10)}

\item{npseudo}{a total number of pseudospots (default = 20000)}

\item{alpha}{loss weights of domain classifier to the source classifier (default = 0.6)}

\item{alpha_lr}{learning rate for the domain classifier (alpha_lr*0.001, default = 5)}

\item{emb_dim}{output size of dimensions for feature extractor (default = 64)}

\item{batch_size}{minibatch size for pseudospots and spatial data during the training (default = 512)}

\item{n_iterations}{iteration number for the adversarial learning (default = 3000)}

\item{init_train_epoch}{iteration number for the pre-training process (default = 10)}

\item{adata_sp}{spatial data (AnnData object) to be used in predicting cell fraction (default: None): count matrix should be non-normalied raw data.}

\item{adata_sc}{single-cell data (AnnData object) to be used in making pseudospots (default: None): count matrix should be non-normalied raw data.}

\item{source.code.dir}{directory of python source code (default = 'pred_cellf_spSeudoMap.py')}

\item{mk_ratio_fix:}{whether to fix the mk_ratio when selecting the}

\item{mk_ratio:}{ratio of number of single-cell markers to virtual pseudotype markers (default = 2)}

\item{pseudo_num_genes:}{number of the virtual markers genes for pseudotype (default = 40)
\itemize{
\item the number is used only when the mk_ratio is not fixed and number of pseudotype markers should be manually provided
}}
}
\value{
spatial data (Seurat object) with predicted cell fraction in metadata (meta.data)
}
\description{
Function to implement spSeudoMap in python
cell type mapping of spatial transcriptomics using unmatched single-cell RNA-seq data
}
\examples{
sp_data_cellf <- pred_cellf_spSeudoMap(sp_data,sc_data_sim,
                                       outdir=output_folder_name,
                                       sp_subset=F,spot.cluster.name='seurat_clusters',
                                       spot.cluster.of.interest=NULL,
                                       env.select='conda',python.install=T,
                                       python_path=NULL,env.name='spSeudoMap',
                                       gpu=TRUE,metadata_celltype='annotation_1',
                                       num_markers=40,mixture_mode='pseudotype',
                                       seed_num=0,
                                       mk_ratio_fix=T, mk_ratio=4,
                                       pseudo_frac_m=pseudo_frac_m, pseudo_frac_std=0.1,
                                       nmix=8,npseudo=npseudo,alpha=0.6,alpha_lr=5,
                                       emb_dim=64,batch_size=512,n_iterations=3000,init_train_epoch=10)

}
